Create big and small headers
Wow did that turn out to be a chore. So hard to figure out the location.
Finally did with globalHistory

Need to figure out a way to change color of svg on hover in link. Going
to have to do that next, after I get my template set up with proper
fontage